Moves mock_case decorator to utils
